GetMapping.. wenn der User admin ist, dann wird er weiter geleitet. wenn nicht dann bekommt er eine error Nachricht

+++ src/main/java/de/hhu/thymeleafsqlseq/controllers/ProductController.java
@@ -2,6 +2,9 @@ package de.hhu.thymeleafsqlseq.controllers;
 
 import de.hhu.thymeleafsqlseq.domain.Product;
 import de.hhu.thymeleafsqlseq.services.ProductService;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -37,9 +40,26 @@ public class ProductController {
     }
 
     @GetMapping("/delete/{id}")
-    public String deleteProduct(@PathVariable("id") Long id){
-        service.deleteProduct(id);
-        return "redirect:/product/list";
+    public String delete(@PathVariable("id") Long id, @AuthenticationPrincipal OAuth2User principal, Model model ){
+        if(principal.getAuthorities().contains("ROLE_ADMIN")){
+            model.addAttribute("id", id);
+            return "redirect:/product/deleteProduct/{id}";
+        }
+        String msg = "you are not an Admin, so you can not delete anything";
+        model.addAttribute("listProducts", service.getAllProducts());
+        model.addAttribute("error", msg);
+        return"listProducts";
+    }
+
+    @Secured("ROLE_ADMIN")
+    @GetMapping("/deleteProduct/{id}")
+    public String deleteProduct(@PathVariable("id") Long id, @AuthenticationPrincipal OAuth2User principal){
+        if(principal.getAuthorities().contains("ROLE_ADMIN")){
+            service.deleteProduct(id);
+            return "redirect:/product/list";
+        }
+
+        return"redirect:/product/list";
     }
 
     @ModelAttribute("product")
